Moved controllers to sub folder and made sure that the client actually works (Newtonsoft.Json) dependency
