+++ src/Augurk.Api/Controllers/V1/TagController.cs
@@ -20,7 +20,7 @@ using System.Web.Http;
 using Augurk.Api.Managers;
 using Augurk.Entities;
 
-namespace Augurk.Api
+namespace Augurk.Api.Controllers.V1
 {
     public class TagController : ApiController
     {
