+++ src/Augurk.CommandLine/packages.config
@@ -1,6 +1,8 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="CommandLineParser" version="1.9.71" targetFramework="net45" />
+  <package id="Microsoft.AspNet.WebApi.Client" version="5.2.3" targetFramework="net45" />
+  <package id="Newtonsoft.Json" version="6.0.4" targetFramework="net45" />
   <package id="SpecFlow" version="1.9.0" targetFramework="net45" />
   <package id="SpecFlow.CustomPlugin" version="1.9.0" targetFramework="net45" />
 </packages>
